Tidy ups
