+++ app/src/main/java/me/banes/chris/tivi/details/epoxymodels/TitleModel.kt
@@ -20,11 +20,11 @@ import android.view.View
 import com.airbnb.epoxy.EpoxyAttribute
 import com.airbnb.epoxy.EpoxyModelClass
 import com.airbnb.epoxy.EpoxyModelWithHolder
-import kotlinx.android.synthetic.main.details_title_item.*
+import kotlinx.android.synthetic.main.view_holder_title.*
 import me.banes.chris.tivi.R
 import me.banes.chris.tivi.ui.epoxymodels.TiviEpoxyHolder
 
-@EpoxyModelClass(layout = R.layout.details_title_item)
+@EpoxyModelClass(layout = R.layout.view_holder_title)
 abstract class TitleModel : EpoxyModelWithHolder<TiviEpoxyHolder>() {
     @EpoxyAttribute var title: String? = null
     @EpoxyAttribute var subtitle: String? = null
