+++ app/src/main/res/layout/view_holder_details_badge.xml
