+++ app/src/main/res/layout/view_holder_empty_state.xml
